Scroll Position
Slowly add scroll position tracker, which will be used to detect where on the page the user is scrolled. This will "activate" sections automatically, which will be helpful for mobile users. Currently, sections are onlu set to active upon mouseEnter. This will solve the other use cases.

Fixed OnMouseOver -> OnMouseEnter to prevent multiple component updates for the Title component. (Overflow)

Start implementing scroll distance logic. This currently has two problems: It is detecting the entire height of the document to be only 899 units, and I think this is due to the scrolling component being separate from the parent component. I must find a way to bind the scroll logic to the outmost wrapper component.

OnMouseLeave for the section wrapper has been deprecated.

Images in the Portfolio section have been temporarily removed, as they are too memory intensive.