(FEATURE/FIX) - Scroll Tracking
This commit will add successful scroll tracking. What this does is keep track of where the position of the scroll is on the client window, and should then automatically "activate" sections when the scroll reaches a certain range.

ScrollSnapView -
* componentDidMount() - This will automatically track the scrolling within this wrapper component element and set the state of the scrollTop value to state.

* calculateScrollDistance - This will take all necessary components of the scroll position -- such as the window height, the scrollPosition from state, and document height (which is the entire element's length), and generate a percentage of completion.

* getDocHeight() - This will find the current document's height in the client window.

SectionWrapper -

This component has been updated to appropriately  detect if a section is active. If it is, it will performt the same functionality that onMouseEnter would provide. (active state is provided by the parent through props)

Welcome/About/Portfolio/Blog/Contact -

All secton components have been updated to receive the props from the parent, which have been generated through ScrollSnapView. When their props meet a certain range of criteria (the scroll percentage), it will send "active" props to the sectionWrapper component that wraps its contents.

index.js -

Passing correct props to the various sections now.

* SetScrollState() - This method is sent to the ScrollSnapView as a prop which is activated through the functionality of the ScrollSnapView component. Eventually, it just sets the state of scrollPosition to the percentage generated within the ScrollSnapView logic. THis is sent to the various section components for activation.

BUGS - currently, this will only function infrequently. It does work when the section is appropriately activated, however it only activates when the position is updated to the correct range for a second time.