Carousel Changes
There are several changes that will alter the carousel's style, as it is now being shared between the Portfolio and About sections.

index.css -- The carousel no longer changes its height dependant on the width of the page, only the height of the page.

Project.js -- Several visual aspects have been changed to adhere to the updated carousel format.

Height has been fixed to a set size that now adheres to the size of the parent carousel slide.

The onClick() function has been deprecated and replaced with an onMouseEnter() function. There is no longer a change of size, and no longer a 'toggling' of the contents within the project component itself -- just an activation upon mouse enter. (this may be deprecated later for the sake of mobile users)

The title and descriptions within the project have been styled to be the correct size to render properly.