Carousel scroll buttons
This commit adds buttons to the carousels, as scrolling left and right proved difficult for most users on desktop and laptop configurations.

ImageCarousel.js - Updated to include buttons.

+ handleScrollRight/Left - These functions are called as onClick functions for the buttons, which in turn call handleScroll()
+ handleScroll() - This take the current element (the carousel), the change to be applied to scroll, and duration of change. In this case, window.innerWidth is the unit to be sent in (either negative or positive) to change the slide, for responsiveness.

+ easeInOutQuad() - Logic for the scroll.

+ The carousel now applies an id to the 'slides' class div dynamically for the buttons to function.

Project.js -

+ handleClassName() - This function has been deprecated, as user testing showed that 'revealing' the content was confusing to users on deskptop. For mobile, this logic was even more confusing.

Portfolio.js - Buttons that switch between carousels for music and software have been updated to utilize the SectionTitle component. This allows for transition effects to be used out of the box.

Button active effects (class name handleMusicBorder() and handlesoftwareBorder()) have been moved inside the title to update styles to be consistent with the parent.

index.js - Removed superfluous console.logs