Bugfix
Arrows seem to not appear correctly on various devices. This commit will test a fix by changing the font family of the arrows to the one that is downloaded correctly and referenced, as the previous was depending on the device already having the font installed.

The hex code for the symbol has also been changed.